Moved Provider classes back into main web package, now package is smaller
