+++ src/main/java/org/concordion/cubano/driver/web/provider/BrowserStackBrowserProvider.java
@@ -1,4 +1,4 @@
-package org.concordion.cubano.driver.web.provider.browserStack;
+package org.concordion.cubano.driver.web.provider;
 
 import org.concordion.cubano.driver.web.provider.RemoteBrowserProvider;
 import org.concordion.cubano.utils.Config;
