+++ src/main/java/org/concordion/cubano/driver/web/provider/RemoteBrowserProvider.java
@@ -8,8 +8,8 @@ import org.apache.http.client.CredentialsProvider;
 import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.concordion.cubano.driver.web.RemoteHttpClientFactory;
-import org.concordion.cubano.driver.web.provider.browserStack.BrowserStackBrowserProvider;
-import org.concordion.cubano.driver.web.provider.sauceLabs.SauceLabsBrowserProvider;
+import org.concordion.cubano.driver.web.provider.BrowserStackBrowserProvider;
+import org.concordion.cubano.driver.web.provider.SauceLabsBrowserProvider;
 import org.concordion.cubano.utils.Config;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.CommandInfo;
